+++ src/main/scala/com/soundsofscala/Instruments/SimpleScala808DrumMachine.scala
@@ -4,7 +4,6 @@ import cats.effect.IO
 import com.soundsofscala.models
 import com.soundsofscala.models.*
 import com.soundsofscala.models.AtomicMusicalEvent.*
-import com.soundsofscala.types.{LookAhead, NextNoteTime, ScheduleWindow, Tempo}
 import org.scalajs.dom
 import org.scalajs.dom.AudioContext
